feat: field mappings
